pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bilibili-api-python>=17.4.0",
    "deepseek>=1.0.0",
    "fastapi>=0.122.0",
    "fastmcp>=2.14.3",
    "matplotlib>=3.10.8",
    "mcp[cli]>=1.22.0",
    "mysql>=0.0.3",
    "mysql-mcp-server>=0.2.2",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "psutil>=7.2.0",
    "scipy>=1.17.0",
    "sympy>=1.14.0",
    "uvicorn>=0.38.0",
]
